refactor: Section package 삭제 및 클래스 이동
