+++ src/main/java/nextstep/subway/line/ui/LineController.java
@@ -4,7 +4,7 @@ import nextstep.subway.exception.InputDataErrorException;
 import nextstep.subway.line.application.LineService;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
-import nextstep.subway.section.domain.dto.SectionRequest;
+import nextstep.subway.line.dto.SectionRequest;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
